Replace numberOfSimulationElements with intensityMapSize outside of Simulation class tree
Simulation::numberOfSimulationElements was previously used as a synonym
to the number of values in the intensity map. However it is different
in the case of SpecularSimulation with beam divergence and could lead
to malfunctioning in the future. To avoid confusion,
numberOfSimulationElements method was placed in the
protected scope of Simulation class. Simulation::intensityMapSize returns
the actual size of intensity map.