Added Read&Write-classes for output data formats; redirected the relevant calls to these classes.
About the reader/writer classes:
Each class contains methods for reading and writing (if applicable) of one format (INT, NumpyTXT, Reflecometry, Tiff). The code for reading and writing is copied from the OutputDataReadStrategy-implementations and from  the OutputDataWriteStrategy-implementations.
For TIFF, the code was already in a separate class (TiffHandler). This has been renamed and reworked a bit (only interface-wise, no implementation changes). Methods from the Tiff strategy have been integrated as well.

The classes have no common polymorphy (no common base class/interface) since it is not necessary in the given implementation. This reduces code & dependencies and eliminates tricks or dummies for non-existing implementations like Reflecometry::write.

Also in the one and only class which uses the readers/writers(IntensityDataIOFactory), polymorphy would not help since there is only one if-else, which would be necessary also in a polymorhic approach.

However, the Reader/Writer classes are designed quite alike, so in case the future brings needs for polymorphy, this can be easily added.

IntensityDataIOFactory:
Code from OutputReader and OutputWriter has been integrated into IntensityDataIOFactory